Refactor playTrack
