+++ packages/common/src/context/comments/commentsContext.tsx
@@ -14,18 +14,23 @@ import {
   useGetCommentsByTrackId,
   useGetCurrentUserId,
   useGetTrackById
-} from '../../api'
+} from '~/api'
+import { useGatedContentAccess } from '~/hooks'
 import {
+  ModalSource,
   ID,
   PaginatedStatus,
   Comment,
   ReplyComment,
   Status,
   UserTrackMetadata
-} from '../../models'
-import { tracksActions } from '../../store/pages/track/lineup/actions'
-import { playerSelectors } from '../../store/player'
-import { Nullable } from '../../utils'
+} from '~/models'
+import { tracksActions } from '~/store/pages/track/lineup/actions'
+import { getLineup } from '~/store/pages/track/selectors'
+import { seek } from '~/store/player/slice'
+import { PurchaseableContentType } from '~/store/purchase-content/types'
+import { usePremiumContentPurchaseModal } from '~/store/ui/modals/premium-content-purchase-modal'
+import { Nullable } from '~/utils'
 
 type CommentSectionProviderProps = {
   entityId: ID
@@ -45,7 +50,7 @@ type CommentSectionContextType = {
   isEntityOwner: boolean
   commentCount: number
   track: UserTrackMetadata
-  playTrack: () => void
+  playTrack: (timestampSeconds?: number) => void
   commentSectionLoading: boolean
   comments: Comment[]
   currentSort: TrackCommentsSortMethodEnum
@@ -95,10 +100,45 @@ export const CommentSectionProvider = (
     }
   )
   const dispatch = useDispatch()
-  const playerUid = useSelector(playerSelectors.getUid) ?? undefined
-  const playTrack = useCallback(() => {
-    dispatch(tracksActions.play(playerUid))
-  }, [dispatch, playerUid])
+
+  const lineup = useSelector(getLineup)
+
+  const { hasStreamAccess } = useGatedContentAccess(track!)
+
+  const { onOpen: openPremiumContentPurchaseModal } =
+    usePremiumContentPurchaseModal()
+
+  const playTrack = useCallback(
+    (timestampSeconds?: number) => {
+      const uid = lineup?.entries?.[0]?.uid
+
+      // If a timestamp is provided, we should seek to that timestamp
+      if (timestampSeconds !== undefined) {
+        // But only if the user has access to the stream
+        if (!hasStreamAccess) {
+          const { track_id: trackId } = track!
+          openPremiumContentPurchaseModal(
+            { contentId: trackId, contentType: PurchaseableContentType.TRACK },
+            {
+              source: ModalSource.Comment
+            }
+          )
+        } else {
+          dispatch(tracksActions.play(uid))
+          setTimeout(() => dispatch(seek({ seconds: timestampSeconds })), 100)
+        }
+      } else {
+        dispatch(tracksActions.play(uid))
+      }
+    },
+    [
+      dispatch,
+      hasStreamAccess,
+      lineup?.entries,
+      openPremiumContentPurchaseModal,
+      track
+    ]
+  )
 
   const commentSectionLoading =
     status === Status.LOADING || status === Status.IDLE
