+++ packages/web/src/components/comments/TimestampLink.tsx
@@ -1,19 +1,5 @@
 import { useCurrentCommentSection } from '@audius/common/context'
-import { useGatedContentAccess } from '@audius/common/hooks'
-import { ModalSource } from '@audius/common/models'
-import {
-  playerActions,
-  PurchaseableContentType,
-  trackPageLineupActions,
-  trackPageSelectors,
-  usePremiumContentPurchaseModal
-} from '@audius/common/store'
 import { formatTrackTimestamp, TextLink, TextLinkProps } from '@audius/harmony'
-import { useDispatch, useSelector } from 'react-redux'
-
-const { tracksActions } = trackPageLineupActions
-const { getLineup } = trackPageSelectors
-const { seek } = playerActions
 
 type TimestampLinkProps = {
   timestampSeconds: number
@@ -21,35 +7,11 @@ type TimestampLinkProps = {
 
 export const TimestampLink = (props: TimestampLinkProps) => {
   const { timestampSeconds, ...other } = props
-
-  const dispatch = useDispatch()
-  const { track } = useCurrentCommentSection()
-  const lineup = useSelector(getLineup)
-  const { track_id: trackId } = track
-
-  const { hasStreamAccess } = useGatedContentAccess(track)
-
-  const uid = lineup?.entries?.[0]?.uid
-  const { onOpen: openPremiumContentPurchaseModal } =
-    usePremiumContentPurchaseModal()
+  const { playTrack } = useCurrentCommentSection()
 
   return (
     <TextLink
-      onClick={() => {
-        if (!hasStreamAccess) {
-          openPremiumContentPurchaseModal(
-            { contentId: trackId, contentType: PurchaseableContentType.TRACK },
-            {
-              source: ModalSource.Comment
-            }
-          )
-        } else {
-          dispatch(tracksActions.play(uid))
-          setTimeout(() => {
-            dispatch(seek({ seconds: timestampSeconds }))
-          })
-        }
-      }}
+      onClick={() => playTrack(timestampSeconds)}
       variant='visible'
       size='s'
       {...other}
